+++ src/main.js
@@ -6,6 +6,8 @@ const { validateEntry } = require('./validate-entry')
 const { parseEntry } = require('./parse-entry')
 const { getEntries } = require('./get-entries')
 const { getVersionById } = require('./get-version-by-id')
+const { getLinks } = require('./get-links')
+const { addLinks } = require('./add-links')
 
 const readFile = utils.promisify(fs.readFile)
 
@@ -23,7 +25,8 @@ exports.main = async function main() {
 
     core.startGroup('Parse data')
     const rawData = await readFile(changelogPath)
-    const versions = getEntries(rawData).map(parseEntry)
+    const linkList = getLinks(rawData)
+    const versions = getEntries(rawData).map(parseEntry).map(addLinks(linkList))
 
     if (validationDepth != 0) {
       const releasedVersions = versions.filter(version => version.status != 'unreleased')